+++ Changelog.md
@@ -4,6 +4,7 @@
   * amq-protocol (now 7.1)
   * cookie-factory (now 0.3)
 * Deprecated `Channel::qos()` in favor of `Channel::set_qos()`
+* `Error::__Nonexhaustive` replaced with `non_exhaustive` attribute
 
 # Version 0.4.2 (2022-01-12)